requirements.txt
glfw==2.3.0
multipledispatch==0.6.0
numpy==1.21.3
PyOpenGL==3.1.5
PyOpenGL-accelerate @ file:///C:/path/PyOpenGL_accelerate-3.1.5-cp39-cp39-win_amd64.whl
pyrr==0.10.3
six==1.16.0
Pillow~=8.4.0